requirements.txt
uvicorn==0.31.0
fastapi==0.115.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
PyMySQL==1.1.1
alembic==1.13.3
bcrypt==4.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
pillow==11.0.0
python-docx==1.1.2
requests==2.32.3
pymongo==4.11
pytz==2025.1
Jinja2==3.1.6
motor==3.7.1